backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
celery==5.3.1
redis==4.6.0
numpy==1.24.4
scrapy==2.9.0
beautifulsoup4==4.12.0
requests==2.31.0
selenium==4.11.2
playwright==1.38.0
django-cors-headers==4.2.0
dj-database-url==2.1.0
gunicorn==21.2.0
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0 
django-filter==25.1
twisted==23.10.0
djoser==2.2.2
lxml==4.9.3